# Add caching of sharding function