@Totktonada 2021-12-30
Since we aim to resolve a performance problem here I would like to see some performance measurements. For example, the following:

* Baseline: just call our own bucket_id function directly (I suggest to extract mpcrc32 from vshard).
* Bad case: call `ddl.bucket_id()` before this patch (name / body).
* Good case: call it after the patch (name / body).

It would be good to commit the microbenchmark right into the repository with a short instruction.

If the good case is not so good, we can look a bit deeper and, say, check, whether implementing a fast path for a function that is placed directly in `_G` will change the situation (here we don't need a loop to extract a nested field with the function). Or, maybe, we'll spot other problematic places. Or, in turn, we can see that the caching does not give anything (I would be surprised, but anyway).